Move avChannel things to its own file
