+++ src/lib/AVCodec/wrappers/AVCodecWrapperInternal.h
@@ -8,6 +8,8 @@
 
 #include <common/InternalTypes.h>
 
+#include "AVChannelInternal.h"
+
 namespace libffmpeg::internal::avcodec
 {
 
@@ -31,78 +33,6 @@ struct AVCodec_56
 
 typedef AVCodec_56 AVCodec_57;
 typedef AVCodec_56 AVCodec_58;
-
-enum AVChannelOrder
-{
-  AV_CHANNEL_ORDER_UNSPEC,
-  AV_CHANNEL_ORDER_NATIVE,
-  AV_CHANNEL_ORDER_CUSTOM,
-  AV_CHANNEL_ORDER_AMBISONIC
-};
-
-enum AVChannel
-{
-  AV_CHAN_NONE = -1,
-  AV_CHAN_FRONT_LEFT,
-  AV_CHAN_FRONT_RIGHT,
-  AV_CHAN_FRONT_CENTER,
-  AV_CHAN_LOW_FREQUENCY,
-  AV_CHAN_BACK_LEFT,
-  AV_CHAN_BACK_RIGHT,
-  AV_CHAN_FRONT_LEFT_OF_CENTER,
-  AV_CHAN_FRONT_RIGHT_OF_CENTER,
-  AV_CHAN_BACK_CENTER,
-  AV_CHAN_SIDE_LEFT,
-  AV_CHAN_SIDE_RIGHT,
-  AV_CHAN_TOP_CENTER,
-  AV_CHAN_TOP_FRONT_LEFT,
-  AV_CHAN_TOP_FRONT_CENTER,
-  AV_CHAN_TOP_FRONT_RIGHT,
-  AV_CHAN_TOP_BACK_LEFT,
-  AV_CHAN_TOP_BACK_CENTER,
-  AV_CHAN_TOP_BACK_RIGHT,
-  AV_CHAN_STEREO_LEFT = 29,
-  AV_CHAN_STEREO_RIGHT,
-  AV_CHAN_WIDE_LEFT,
-  AV_CHAN_WIDE_RIGHT,
-  AV_CHAN_SURROUND_DIRECT_LEFT,
-  AV_CHAN_SURROUND_DIRECT_RIGHT,
-  AV_CHAN_LOW_FREQUENCY_2,
-  AV_CHAN_TOP_SIDE_LEFT,
-  AV_CHAN_TOP_SIDE_RIGHT,
-  AV_CHAN_BOTTOM_FRONT_CENTER,
-  AV_CHAN_BOTTOM_FRONT_LEFT,
-  AV_CHAN_BOTTOM_FRONT_RIGHT,
-  AV_CHAN_SIDE_SURROUND_LEFT,  ///<  +90 degrees, Lss, SiL
-  AV_CHAN_SIDE_SURROUND_RIGHT, ///<  -90 degrees, Rss, SiR
-  AV_CHAN_TOP_SURROUND_LEFT,   ///< +110 degrees, Lvs, TpLS
-  AV_CHAN_TOP_SURROUND_RIGHT,  ///< -110 degrees, Rvs, TpRS
-  AV_CHAN_BINAURAL_LEFT = 61,
-  AV_CHAN_BINAURAL_RIGHT,
-  AV_CHAN_UNUSED         = 0x200,
-  AV_CHAN_UNKNOWN        = 0x300,
-  AV_CHAN_AMBISONIC_BASE = 0x400,
-  AV_CHAN_AMBISONIC_END  = 0x7ff
-};
-
-struct AVChannelCustom
-{
-  AVChannel id{};
-  char      name[16]{};
-  void     *opaque{};
-};
-
-struct AVChannelLayout
-{
-  AVChannelOrder order{};
-  int            nb_channels{};
-  union
-  {
-    uint64_t         mask;
-    AVChannelCustom *map;
-  } u{};
-  void *opaque{};
-};
 
 struct AVCodec_59
 {
